[BUGFIX] iwlwifi: never create empty regdomain
On older firmware, such as for 7265, querying the regdomain while
the NIC is in (hard) rf-kill causes the firmware to return a list
of channels that are all disabled. This in turn causes the driver
to create an empty regdomain, causing cfg80211 to throw a warning
and fail, this failure bubbles up, the driver doesn't initialize,
and ultimately wifi cannot be used.

We didn't have this problem until fairly recently because there's
now a query of the regdomain during initialization, but that's a
bugfix for an issue where cfg80211 will (rightfully) complain in
a scenario where somebody tries to query the regdomain for a NIC,
the driver has advertised custom regdomain support but there's no
regdomain installed.

To avoid regression on that original issue, add a dummy rule to
the regdomain if it ended up being empty - this dummy rule has no
usable overlap with any channel, but is otherwise valid. When the
rf-kill is removed and the interface is brought up, we'll update
the regdomain from the firmware, now with valid data.

type=bugfix
ticket=jira:WIFI-122207
fixes=I7d8d5c895bc467efbf81ea055dde366ea01cced1

Change-Id: I498e401a45d080522d9d345f09a5b209882fa695
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/320003
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Luciano Coelho <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 75447b26b43a2744a360e124f77465f58f25f4e0