wifi: iwlwifi: limit EHT capabilities based on PCIe link speed
If a discrete NIC is connected to a PCIe link hat isn't at least
Gen3 (8.0 GT/s), then we cannot sustain 320 MHz traffic, so remove
that from EHT capabilities in that case.

While at it, also move setting 320 MHz beamformee to the right
place in the code so it's not set while not supporting 320 MHz.

type=feature
ticket=jira:WIFI-233736

Change-Id: Id4120c161fb7df6dedc70d5f3e3829e9117b8cb1
Signed-off-by: Johannes Berg <johannes.berg@intel.com>x-iwlwifi-stack-dev: b8feb22b1e6049743fe82510bc22f3c847e51a4e