[NOUPSTREAM] wifi: iwlwifi: mvm: increase debug level when we fail to associate
Chrome users report a high number of "Not associated and the time event
is over already..." errors.
By looking at the logs it seems that we can't get the authentication to
go out, or at least, we don't hear the authentication response.
Add more logs to debug this area specifically.
This is a temporary patch to debug this specific issue that should be
shipped to the users and the code is written in such a way that it'll be
easy to ship it to Chrome users only: under an ifdef.
This is also the reason the copyright is not changed: it makes little
sense to change the copyright for a patch that will be reverted.

type=feature
ticket=jira:WIFI-290200

Change-Id: I09d5a0983adf88c426b234eecf05a3b2bd768a1d
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/46682
Reviewed-by: Greenman, Gregory <gregory.greenman@intel.com>
tested: Greenman, Gregory <gregory.greenman@intel.com>
x-iwlwifi-stack-dev: db4cbef6ee59dcfb15f7dd75d54424ed1b8920cd