wifi: iwlwifi: mvm: limit EHT 320 MHz MCS for STEP URM
If the STEP (the interface between MAC and PHY) is in URM
(a lower speed mode) then we cannot use 320 MHz MCS > 9.
Therefore, limit the MCS in our capabilities in this case.
Note that this also limits the TX/rate scaling since that
takes both TX and RX capabilities into account.

type=feature
ticket=none

Change-Id: Id5efbb71d45da02c8c4e211d20396637ddd44da8
Signed-off-by: Johannes Berg <johannes.berg@intel.com>x-iwlwifi-stack-dev: 93af5bb084b610c99e9913e446685a7ebe729b1d